+++ src/pages/AdminProductTypePage/AdminProductTypePage.js
@@ -1,10 +1,10 @@
 import React from "react";
 // Variable Width
 import { Div, Dropdown, Anchor } from "atomize";
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
+import NavBar from "../../Components/NavBar";
+import Footer from "../../Components/Footer";
 
-const StoreListOption = (
+const ProductListOption = (
   <Div p={{ x: "1rem", y: "0.5rem" }}>
     {["生鮮雜貨", "冷藏肉品", "零食", "飲品", "其他"].map((name, index) => (
       <Anchor d="block" p={{ y: "0.25rem" }}>
@@ -13,7 +13,7 @@ const StoreListOption = (
     ))}
   </Div>
 );
-class StoreTypeFilter extends React.Component {
+class ProductTypeFilter extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -27,14 +27,14 @@ class StoreTypeFilter extends React.Component {
         w="fit-content"
         isOpen={showDropdown}
         onClick={() => this.setState({ showDropdown: !showDropdown })}
-        menu={StoreListOption}
+        menu={ProductListOption}
       >
-        商家類別
+        商品種類
       </Dropdown>
     );
   }
 }
-const StoreList = () => {
+const ProductList = () => {
   return (
     <Div
       border="1px solid"
@@ -45,25 +45,25 @@ const StoreList = () => {
       justify="space-between"
     >
       {" "}
-      <Div transform="translateY(25%)">1. 商家編號：5298974923</Div>
-      <StoreTypeFilter />
+      <Div transform="translateY(25%)">1. 商品編號：5298974923</Div>
+      <ProductTypeFilter />
     </Div>
   );
 };
 
-const AdminStoreTypePage = () => {
+const AdminProductTypePage = () => {
   return (
     <Div>
       <NavBar />
       <Div m={{ l: "5rem", r: "5rem" }}>
-        <StoreTypeFilter />
-        <StoreList></StoreList>
-        <StoreList />
-        <StoreList />
+        <ProductTypeFilter />
+        <ProductList></ProductList>
+        <ProductList />
+        <ProductList />
       </Div>
       <Footer />
     </Div>
   );
 };
 
-export default AdminStoreTypePage;
+export default AdminProductTypePage;
