+++ src/pages/AdminStoreTypePage/AdminStoreTypePage.js
@@ -1,10 +1,10 @@
 import React from "react";
 // Variable Width
 import { Div, Dropdown, Anchor } from "atomize";
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
+import NavBar from "../../Components/NavBar";
+import Footer from "../../Components/Footer";
 
-const ProductListOption = (
+const StoreListOption = (
   <Div p={{ x: "1rem", y: "0.5rem" }}>
     {["生鮮雜貨", "冷藏肉品", "零食", "飲品", "其他"].map((name, index) => (
       <Anchor d="block" p={{ y: "0.25rem" }}>
@@ -13,7 +13,7 @@ const ProductListOption = (
     ))}
   </Div>
 );
-class ProductTypeFilter extends React.Component {
+class StoreTypeFilter extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -27,14 +27,14 @@ class ProductTypeFilter extends React.Component {
         w="fit-content"
         isOpen={showDropdown}
         onClick={() => this.setState({ showDropdown: !showDropdown })}
-        menu={ProductListOption}
+        menu={StoreListOption}
       >
-        商品種類
+        商家類別
       </Dropdown>
     );
   }
 }
-const ProductList = () => {
+const StoreList = () => {
   return (
     <Div
       border="1px solid"
@@ -45,25 +45,25 @@ const ProductList = () => {
       justify="space-between"
     >
       {" "}
-      <Div transform="translateY(25%)">1. 商品編號：5298974923</Div>
-      <ProductTypeFilter />
+      <Div transform="translateY(25%)">1. 商家編號：5298974923</Div>
+      <StoreTypeFilter />
     </Div>
   );
 };
 
-const AdminProductTypePage = () => {
+const AdminStoreTypePage = () => {
   return (
     <Div>
       <NavBar />
       <Div m={{ l: "5rem", r: "5rem" }}>
-        <ProductTypeFilter />
-        <ProductList></ProductList>
-        <ProductList />
-        <ProductList />
+        <StoreTypeFilter />
+        <StoreList></StoreList>
+        <StoreList />
+        <StoreList />
       </Div>
       <Footer />
     </Div>
   );
 };
 
-export default AdminProductTypePage;
+export default AdminStoreTypePage;
