+++ src/pages/FAQPage/index.js
@@ -1,0 +1,1 @@
+export { default } from "./FAQPage";
