+++ src/pages/SignIn/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SignIn";
